Reorganized code into different files; namespaces are preserved
